requirements.txt
# LegacyLens Python Dependencies

# --- Core Framework ---
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.5.0

# --- Vector Database ---
qdrant-client>=1.7.0

# --- Embeddings ---
voyageai>=0.3.0

# --- LLM ---
openai>=1.10.0

# --- Re-ranking ---
cohere>=5.0.0

# --- Tokenization ---
tiktoken>=0.6.0

# --- Encoding Detection ---
chardet>=5.2.0

# --- CLI ---
click>=8.1.0
rich>=13.7.0
httpx>=0.27.0

# --- Testing ---
pytest>=8.0.0
pytest-asyncio>=0.23.0

# --- Linting ---
ruff>=0.2.0
